Option to close instead of deleting buffers
The current behavior uses `:bd` to close a diff window after it is displayed on
the screen, which causes any other windows displaying that file to be closed.

The option g:DirDiffBufferDelete allows to use `:q` instead. In this case the
`:diffoff` is also necessary to avoid subsequent diffs to be displayed
incorrectly.